+++ src/app/cart/cart.module.ts
@@ -3,8 +3,6 @@ import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-list/cart-item/cart-item.component';
 
-
-
 @NgModule({
   declarations: [
     CartListComponent,