+++ src/app/cart/models/cart.model.ts
@@ -1,6 +1,7 @@
 export default class CartItemModel {
   constructor(
     public name: string,
+    public price: number,
     public quantity: number,
     public amount: number,
   ) { }