+++ src/app/cart/services/cart-push.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CartPushService } from './cart-push.service';
+
+describe('CartPushService', () => {
+  let service: CartPushService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CartPushService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});