+++ src/app/shared/directives/drawborder.directive.spec.ts
@@ -1,0 +1,8 @@
+import { DrawborderDirective } from './drawborder.directive';
+
+describe('DrawborderDirective', () => {
+  it('should create an instance', () => {
+    const directive = new DrawborderDirective();
+    expect(directive).toBeTruthy();
+  });
+});